ERF-7660: Vue 3 kompatibel
