+++ src/types/vue-multiselect.d.ts
@@ -1,0 +1,1 @@
+declare module 'vue-multiselect';
