+++ vue.config.js
@@ -40,6 +40,10 @@ const options = {
 if (process.env.VG_VUE_TRANSPILE === 'true') {
   console.log('⚙️ Transpiling Vue applikation ...');
   options.css.extract = false;
+
+  options.configureWebpack.externals = {
+    vue: 'Vue'
+  }
 }
 
 module.exports = options;
